requirements.txt
python-dotenv
pandas
matplotlib
numpy 
pymongo
certifi
scikit-learn
pymongo[srv]==3.12
# -e .